kak vk

+++ Lab7Server/src/DataBaseCommunication.java
@@ -49,28 +49,34 @@ public class DataBaseCommunication {
             if(message.getTypeOfOperation()==Message.change) {
                 statement.execute("update normalhuman set name = '" + NewData.get(0).getName() + "', age = " + NewData.get(0).getAge() + ", troubleswiththelaw = " + NewData.get(0).getTroublesWithTheLaw() + " where id = " + NewData.get(0).getId() + ";");
                 statement.execute("delete from thoughts where id=" + NewData.get(0).getId()+";");
-                StringBuilder sql = new StringBuilder();
-                sql.append("insert into thoughts values('"+NewData.get(0).getThoughts(0) +"', "+ NewData.get(0).getId() +" )");
-                for(int i=1;i<NewData.get(0).countOfThoughts();i++){
-                    sql.append(",('"+NewData.get(0).getThoughts(i) +"', "+ NewData.get(0).getId() +" )");
+                if(NewData.get(0).getThoughtsCount()!=0) {
+                    StringBuilder sql = new StringBuilder();
+                    sql.append("insert into thoughts values('" + NewData.get(0).getThoughts(0) + "', " + NewData.get(0).getId() + " )");
+                    for (int i = 1; i < NewData.get(0).countOfThoughts(); i++) {
+                        sql.append(",('" + NewData.get(0).getThoughts(i) + "', " + NewData.get(0).getId() + " )");
+                    }
+                    sql.append(";");
+                    System.out.println(sql);
+                    statement.execute(sql.toString());
                 }
-                sql.append(";");
-                System.out.println(sql);
-                statement.execute(sql.toString());
             }
             else if(message.getTypeOfOperation()==Message.add) {
                 statement.execute("insert into normalhuman values (" + NewData.get(0).getId() + ", '" + NewData.get(0).getName() + "'," + NewData.get(0).getAge() + "," + NewData.get(0).getTroublesWithTheLaw() + ");");
-                StringBuilder sql = new StringBuilder();
-                sql.append("insert into thoughts values('"+NewData.get(0).getThoughts(0) +"', "+ NewData.get(0).getId() +" )");
-                for(int i=1;i<NewData.get(0).countOfThoughts();i++){
-                    sql.append(",('"+NewData.get(0).getThoughts(i) +"', "+ NewData.get(0).getId() +" )");
+                if(NewData.get(0).getThoughtsCount()!=0) {
+                    StringBuilder sql = new StringBuilder();
+                    sql.append("insert into thoughts values('" + NewData.get(0).getThoughts(0) + "', " + NewData.get(0).getId() + " )");
+                    for (int i = 1; i < NewData.get(0).countOfThoughts(); i++) {
+                        sql.append(",('" + NewData.get(0).getThoughts(i) + "', " + NewData.get(0).getId() + " )");
+                    }
+                    sql.append(";");
+                    statement.execute(sql.toString());
                 }
-                sql.append(";");
-                statement.execute(sql.toString());
             }
             else if(message.getTypeOfOperation()==Message.delete) {
                 statement.execute("delete from thoughts where id=" + NewData.get(0).getId() + ";");
-                statement.execute("delete from normalhuman where id = " + NewData.get(0).getId() + ";");
+                if(NewData.get(0).getThoughtsCount()!=0) {
+                    statement.execute("delete from normalhuman where id = " + NewData.get(0).getId() + ";");
+                }
             }
             statement.close();
             connection.close();
